Update advanced.rst
I think the example got mixed up a bit and some of the words were placed incorrectly. IEventHandler<in T> would be able to handle T if it was a more derived class, as expected in the registration example.

Before:
which makes *IEventHandler<SendFlowerToMovedCustomer>* assignable to *IEventHandler<CustomerMovedAbroadEvent>*
(if SendFlowerToMovedCustomer was supposed to be CustomerMovedEvent, I don't think this would still be a valid scenario [if I am understanding this all correctly])

After:
which makes *IEventHandler<CustomerMovedAbroadEvent>* assignable to *IEventHandler<CustomerMovedEvent>*
(The more abstract handler could handle the more derived handler, but not the other way around)

Please correct me if my I am misunderstanding this. Thanks