Improvement to web forms documentation
Based on the possible breaking changes in v5.0 considering the ability
to resolve unregistered types, the Web Forms documentation has been
updated to keep working OOTB when this change is introduced. By
introducing the change in documentation at this time, projects that
apply this integration advice at this  time will benefit when updating
to v5.0 when that is released.