Textual improvements
Reformatted code blocks.
Renamed "automatic constructor injection" to "auto-wiring" which is
in-line with the book's terminology.
Renamed "batch-registration" to "auto-registration", inline with the
book.
Added a reference to the book's Composition Root article.
Improved the advise of using Auto-Wiring to using Auto-Wiring -only- for
applicatoin components, not to third-party components.